hard-coded TPS limit, does not work due to the fact that the slimevr ecosystem does not support any TPS limiting whatso-ever.
specifically, it does work, although when doing any kind of movements (especially letting them rest for a while),

let us say there's two points, A and B.
of course with low TPS your movements will be a bit delayed, although this is for the user to decide what experience they want without spending more money on a router.

if you move a body part from point A to point B, the in real life tracker is at point B, however because the server receives late data, it will take time to reflect that the tracker moved to point B, and hence when your body part/tracker is still, it will then slowly move to point B, so

when you did that movement, the actual tracker position server-side would be somewhere between point A and point B, or close, depending on how fast the movement was.

This is not a problem of limiting the TPS but rather, the slimevr ecosystem NEVER supported limiting TPS in the first place.