TODO: Finish adding configuration where you can save/read JSON files. TODO: Also figure out what to do with I2C scanning. Maybe begin implementing magnometers?
I really like this way of initializing the sensors/IMUs. I think it would be much more efficient to control what sensors are initialized by defines (and thus you won't be loading a lot of libraries for different IMUs that you don't have or won't be using) and you can save space this way.

Before, all libraries would've been loaded and initialized, in the official SlimeVR firmware code, to account for any sensors that may be connected.

Perhaps in the future, this code will be ergonomic enough to support dynamic IMU initialization. (it may be possible anyway by I2C scanning, correct me if i'm wrong, I am very new to this.)

Because of the fact that every IMU class/struct inherits from a base struct, and can override functions, you can have a completely dynamic IMU that functions different from every other defined IMU/sensor, and yet AS LONG AS THERE IS FUNCTIONS THAT EXIST IN THE BASE STRUCT, you can still use the same functions to get around the same behavior of the IMU.

So therefore, dynamic IMU initialization is possible.