1. position estimation is working. dead reckoning is working (albeit with fast accumulative drift).
2. wifi works now. (should connect to last used access point.)

3. calibration works now. (all it currently does is sample biases/error from the imu accelerometer to maybe: A. estimate position based off of accelerometer alone with more advanced methods rather than just using a filter all-in-all or B. use a filter and use these biases, which probably won't matter anyway because of the drift.)

You need to turn the tracker upside-down for calibration to initiate.
Also, I personally vote for A for these biases to be utilized.

If drift happens anyways with B, why not just go for A and go for the most lite, and more efficient/practical approach? With B, you're practically doing matrix multiplication for something like the Kalman Filter, which can also add a lot more memory usage, and slow down performance of the board.

And you can use more methods/advanced methods with A to lessen drift even more by just utilizing what we have at hand. I.e, you can reset position/velocity to 0 when stationary to remove accumulative drift thus far. **that stock slimevr doesnt have**. this is practically ZUPT.

and, to negate the double integration (where drift can quadratically increase) issue, we can simply multiply these thresholds over time, based off of the sampling method in particular.

So like, say for example the accelerometer had bias/error of 1 on every axis.
A = (current acceleration-acceleration threshold)
before sampling, we initialize `acceleration threshold` with the value of the current acceleration of the current imu. (so from the start `acceleration threshold` is something like 0,0,0 so we initialize it with the current acceleration value which is where the imu is currently facing, so the result would be something like 0,0,-1 if the imu is facing downwards, therefore we can catch the biases/errors with this)

during sampling, we catch any and all biases, and multiply the acceleration threshold by the max coefficient of A. (this is not precise although it also accounts for any drift in other axis' all at once rather than going for precision and checking the axis' one by one.)

so now, during movement:
double integration is basically the error is amplified quadratically **over time.**

so to combat this, all we have to do, during sampling, we can estimate just how much is accumulated overtime, and how long for it to reach that value.

and then whenever the tracker moves, we can multiply the thresholds using these values to negate a lot of drift/error.

This **seems** better than B simply because we are more free to do what we wish than just using a library for this, and we can reach our goal more simpler than doing something like matrix multiplication (kalman filter). It's pretty straightforward.

3. task manager exists now. it's just a way to avoid making .h and .cpp files respectively for one class as a whole. in stock slimevr firmware, a single class can be represented as both (1 .h file, and 1. cpp file.) the **h file just plans everything out basically, and the cpp file does the work, for simplicity.**

what i wanted this to do is basically, allow us to simply make 1 .h file instead of making 2 separate files, to do one task. (and to also work with the limitations of the esp8266 board)

we can also store data in this same task, and the same task has configuration to be ran once, or ran at an interval, or by default run every `loop` call.

which makes this better than slimevr's current way of doing things.